+++ ipwhois/asn.py
@@ -169,16 +169,6 @@ class IPASN:
 
         return ret
 
-    def _parse_fields_dns(self, *args, **kwargs):
-        """
-        Deprecated. This will be removed in a future release.
-        """
-
-        from warnings import warn
-        warn('IPASN._parse_fields_dns() has been deprecated and will be '
-             'removed. You should now use IPASN.parse_fields_dns().')
-        return self.parse_fields_dns(*args, **kwargs)
-
     def parse_fields_verbose_dns(self, response):
         """
         The function for parsing ASN fields from a verbose dns response.
@@ -293,16 +283,6 @@ class IPASN:
 
         return ret
 
-    def _parse_fields_whois(self, *args, **kwargs):
-        """
-        Deprecated. This will be removed in a future release.
-        """
-
-        from warnings import warn
-        warn('IPASN._parse_fields_whois() has been deprecated and will be '
-             'removed. You should now use IPASN.parse_fields_whois().')
-        return self.parse_fields_whois(*args, **kwargs)
-
     def parse_fields_http(self, response, extra_org_map=None):
         """
         The function for parsing ASN fields from a http response.
@@ -403,19 +383,8 @@ class IPASN:
 
         return asn_data
 
-    def _parse_fields_http(self, *args, **kwargs):
-        """
-        Deprecated. This will be removed in a future release.
-        """
-
-        from warnings import warn
-        warn('IPASN._parse_fields_http() has been deprecated and will be '
-             'removed. You should now use IPASN.parse_fields_http().')
-        return self.parse_fields_http(*args, **kwargs)
-
-    def lookup(self, inc_raw=False, retry_count=3, asn_alts=None,
-               extra_org_map=None, asn_methods=None,
-               get_asn_description=True):
+    def lookup(self, inc_raw=False, retry_count=3, extra_org_map=None,
+               asn_methods=None, get_asn_description=True):
         """
         The wrapper function for retrieving and parsing ASN information for an
         IP address.
@@ -426,10 +395,6 @@ class IPASN:
             retry_count (:obj:`int`): The number of times to retry in case
                 socket errors, timeouts, connection resets, etc. are
                 encountered. Defaults to 3.
-            asn_alts (:obj:`list`): Additional lookup types to attempt if the
-                ASN dns lookup fails. Allow permutations must be enabled.
-                Defaults to all ['whois', 'http']. *WARNING* deprecated in
-                favor of new argument asn_methods. Defaults to None.
             extra_org_map (:obj:`dict`): Mapping org handles to RIRs. This is
                 for limited cases where ARIN REST (ASN fallback HTTP lookup)
                 does not show an RIR as the org handle e.g., DNIC (which is
@@ -466,17 +431,7 @@ class IPASN:
 
         if asn_methods is None:
 
-            if asn_alts is None:
-
-                lookups = ['dns', 'whois', 'http']
-
-            else:
-
-                from warnings import warn
-                warn('IPASN.lookup() asn_alts argument has been deprecated '
-                     'and will be removed. You should now use the asn_methods '
-                     'argument.')
-                lookups = ['dns'] + asn_alts
+            lookups = ['dns', 'whois', 'http']
 
         else:
 
@@ -492,8 +447,7 @@ class IPASN:
         dns_success = False
         for index, lookup_method in enumerate(lookups):
 
-            if index > 0 and not asn_methods and not (
-                    self._net.allow_permutations):
+            if index > 0 and not asn_methods:
 
                 raise ASNRegistryError('ASN registry lookup failed. '
                                        'Permutations not allowed.')
@@ -706,16 +660,6 @@ class ASNOrigin:
 
         return ret
 
-    def _parse_fields(self, *args, **kwargs):
-        """
-        Deprecated. This will be removed in a future release.
-        """
-
-        from warnings import warn
-        warn('ASNOrigin._parse_fields() has been deprecated and will be '
-             'removed. You should now use ASNOrigin.parse_fields().')
-        return self.parse_fields(*args, **kwargs)
-
     def get_nets_radb(self, response, is_http=False):
         """
         The function for parsing network blocks from ASN origin data.
@@ -769,18 +713,8 @@ class ASNOrigin:
 
         return nets
 
-    def _get_nets_radb(self, *args, **kwargs):
-        """
-        Deprecated. This will be removed in a future release.
-        """
-
-        from warnings import warn
-        warn('ASNOrigin._get_nets_radb() has been deprecated and will be '
-             'removed. You should now use ASNOrigin.get_nets_radb().')
-        return self.get_nets_radb(*args, **kwargs)
-
     def lookup(self, asn=None, inc_raw=False, retry_count=3, response=None,
-               field_list=None, asn_alts=None, asn_methods=None):
+               field_list=None, asn_methods=None):
         """
         The function for retrieving and parsing ASN origin whois information
         via port 43/tcp (WHOIS).
@@ -797,9 +731,6 @@ class ASNOrigin:
             field_list (:obj:`list`): If provided, fields to parse:
                 ['description', 'maintainer', 'updated', 'source']
                 If None, defaults to all.
-            asn_alts (:obj:`list`): Additional lookup types to attempt if the
-                ASN whois lookup fails. If None, defaults to all ['http'].
-                *WARNING* deprecated in favor of new argument asn_methods.
             asn_methods (:obj:`list`): ASN lookup types to attempt, in order.
                 If None, defaults to all ['whois', 'http'].
 
@@ -828,17 +759,7 @@ class ASNOrigin:
 
         if asn_methods is None:
 
-            if asn_alts is None:
-
-                lookups = ['whois', 'http']
-
-            else:
-
-                from warnings import warn
-                warn('ASNOrigin.lookup() asn_alts argument has been deprecated'
-                     ' and will be removed. You should now use the asn_methods'
-                     ' argument.')
-                lookups = ['whois'] + asn_alts
+            lookups = ['whois', 'http']
 
         else: